// src/types/GithubUser.ts
export type GithubUser = {
  login: string;
  id: number;
  avatar_url: string;
  html_url: string;

  name: string | null;
  company: string | null;
  blog: string;
  location: string | null;
  bio: string | null;

  public_repos: number;
  followers: number;
  following: number;

  created_at: string;
  updated_at: string;
};